Fix #4. maxima/minima of a constant value array has no peaks
Previously, we didn't check if the less-than comparison between 2
elements was false because they were equal, this erroneously added peaks
in every window due to the default of `peak = true`.

Added checks to avoid comparing an element with itself, and checks to
treat equality between different elements as not a peak.