Change includebounds => strictbounds; fix missing/NaN behavior
The previous behavior of missing/NaN behavior with `includebounds` was not
logically consistent, and could not be with the original definition of
`includebounds`. To ensure that the behavior of missing/NaNs was
logically consistent with bounding behavior, `includebounds` was changed
to `strictbounds`. `strictbounds` has the same intent, but broader
application. See updated docstrings for more info.