Added note about FWHM
I have been looking for a way to compute the full width half maximum of functions and vectors for some time (even deciding to make a package for it!) before I realized that Peaks.jl actually offers this functionality. I have now added a note to make that more discoverable. Feel free to move the note and change the wording - it is mostly intended for google to return the package.